+++ src/web/schemas/poem_forms.py
@@ -3,7 +3,7 @@ from typing import Annotated, Self, Union
 from pydantic import BaseModel, Field, model_validator
 
 from src import api as domain
-from .weights import Weights
+from src.web.schemas._util import to_kebab
 
 type Rhythm = list[Union[int, str, None]]
 
@@ -46,6 +46,27 @@ class Prosody(BaseModel):
     def as_domain(self) -> domain.Prosody:
         """Converts the Schema to Domain's model"""
         return domain.Prosody(self.pattern, self.rhythm)
+
+
+class Weights(BaseModel):
+    vocal_harmony: float = Field(default=1, ge=0, le=1)
+    accentuation: float = Field(default=1, ge=0, le=1)
+    tonic_position: float = Field(default=1, ge=0, le=1)
+    internal_rhyme: float = Field(default=1, ge=0, le=1)
+    rhythmic_structure: float = Field(default=1, ge=0, le=1)
+
+    def as_domain(self) -> domain.Weights:
+        return domain.Weights(
+            accentuation=self.accentuation,
+            internal_rhyme=self.internal_rhyme,
+            rhythmic_structure=self.rhythmic_structure,
+            tonic_position=self.tonic_position,
+            vocal_harmony=self.vocal_harmony,
+        )
+
+    class Config:
+        alias_generator = to_kebab
+        populate_by_name = True
 
 
 class PropoeForms(BaseModel):
